Extract DateRangeEditor to a separate demo since it only works with Qt
